requirements.txt
pygame-ce~=2.5.2
PyTMX~=3.32